Move GitHub dispatch tests out of Jira defaults test module
